requirements/base.txt
Django==3.1
djangorestframework==3.11.0
django-environ==0.4.5
django-extensions==2.2.9
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.2.1
ipython==7.15.0
requests==2.23.0
django-debug-toolbar==2.2
flake8==7.0.0
